更新代码逻辑

+++ Source/WHFramework/Private/Scene/Components/WorldTimerComponent.cpp
@@ -31,7 +31,7 @@ void UWorldTimerComponent::UpdateTimer()
 {
 	TimeSeconds = FMath::TruncToInt(UGameplayStatics::GetTimeSeconds(this));
 
-	double RemainSeconds = 0;
+	float RemainSeconds = 0;
 	CurrentDay = UKismetMathLibrary::FMod(TimeSeconds, SecondsOfDay, RemainSeconds);
 
 	UpdateLight(RemainSeconds/ (SecondsOfDay / 24));
